Refactors Utils.py module
