+++ pyunfold/Mix.py
@@ -7,7 +7,7 @@
 from itertools import product
 import numpy as np
 from .CovMatrix import CovarianceMatrix
-from .Utils import safe_inverse, none_to_empty_list
+from .utils import safe_inverse, none_to_empty_list
 
 
 class Mixer(object):
